more refactors to token CRUD
